Arduino System
This Arduino system detects if your heart rate is below 120 BPM. If so, it activates a motor that functions as a piston, punching the user. This is a proof of concept, and if we were to implement this for real, we would add a Bluetooth connector and other components to make the software and hardware systems compatible with each other.